fix(tests): switch GPU model tests to bf16 for flash attention compatibility
PR #3057 removed sdpa/eager attention, leaving only flash_attention_2/3/4
which require fp16/bf16. Tests still used default_dtype(torch.float32),
causing 27 GPU unit test failures on main.

Changes:
- Switch default_dtype(torch.float32) -> default_dtype(torch.bfloat16) in
  all affected model comparison tests (llama, glm4_moe, nemotron_h,
  nemotron_h_kl, qwen3_5_moe, qwen3_5_moe_vlm, qwen3_moe)
- Skip afmoe tests: HF afmoe model does not support flash attention 2
- test_fused_lm_head::test_full_model_fused_vs_vanilla: switch to bf16
  model dtype + loosen tolerances for half-precision
- Fix view(-1) -> reshape(-1) in sequence.py: FA3 code path calls
  get_cu_seqlens_from_position_ids with non-contiguous position_ids
  (from .expand()), which .view() rejects
- Loosen nemotron_h_kl fused vs vanilla tolerance from 1e-3 to 1e-2 for bf16

+++ tests/unit/train/models/test_glm4_moe.py
@@ -33,7 +33,7 @@ def get_model_pairs() -> tuple[HFGlm4MoeForCausalLM, PrimeRLGlm4MoeForCausalLM]:
     # But the grad seems to be off in attn because of precision
     # hf_config._attn_implementation = "flash_attention_2"
     hf_config._attn_implementation = "flash_attention_2"
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         hf_model = HFGlm4MoeForCausalLM._from_config(hf_config)
         prime_model = PrimeRLGlm4MoeForCausalLM._from_config(hf_config)
     with torch.no_grad():
@@ -59,7 +59,7 @@ def test_glm4_moe_attn_only() -> None:
     for layer in prime_model.model.layers:
         layer.mlp = _IdentityMLP()
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, hf_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
 
@@ -87,7 +87,7 @@ def test_glm4_moe_mlp_only() -> None:
     for layer in prime_model.model.layers:
         layer.self_attn.forward = foo
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, hf_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
 
@@ -107,7 +107,7 @@ def test_glm4_moe_mlp_only() -> None:
 def test_glm4_moe() -> None:
     hf_model, prime_model = get_model_pairs()
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, hf_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
