+++ tests/unit/train/models/test_llama.py
@@ -26,7 +26,7 @@ def get_model_pairs():
         mlp_bias=False,
     )
     hf_config._attn_implementation = "flash_attention_2"
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         hf_model = HFLlamaForCausalLM._from_config(hf_config)
         prime_model = PrimeRLLlamaForCausalLM._from_config(hf_config)
     with torch.no_grad():
@@ -47,7 +47,7 @@ def test_llama_attn_only():
     for layer in prime_model.model.layers:
         layer.mlp = nn.Identity()
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, hf_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
 
@@ -75,7 +75,7 @@ def test_llama_mlp_only():
     for layer in prime_model.model.layers:
         layer.self_attn.forward = foo
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, hf_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
 
@@ -95,7 +95,7 @@ def test_llama_mlp_only():
 def test_llama():
     hf_model, prime_model = get_model_pairs()
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         input_ids = torch.randint(0, hf_model.config.vocab_size, (1, 100))
         position_ids = torch.arange(1, 101).unsqueeze(0)
 
@@ -111,7 +111,7 @@ def test_llama():
     grad_diff = hf_model.model.embed_tokens.weight.grad - prime_model.model.embed_tokens.weight.grad
     assert torch.allclose(grad_diff, torch.zeros_like(grad_diff), atol=2), f"Max grad diff: {grad_diff.abs().max()}"
 
-    with torch.device("cuda"), default_dtype(torch.float32):
+    with torch.device("cuda"), default_dtype(torch.bfloat16):
         hf_from_prime_model = HFLlamaForCausalLM._from_config(hf_model.config)
         converted_state_dict = prime_model.convert_to_hf(prime_model.state_dict())
         hf_from_prime_model.load_state_dict(converted_state_dict)
